app/requirements.txt
flask
numpy
galois
pillow
scipy